fix for static builds
