+++ src/devices/messages/ControlBoardMsgs/idl_generated_code/ControlBoardMsgs_index.txt
@@ -2,12 +2,12 @@ return_isJointBraked.h
 return_isJointBraked.cpp
 return_getAutoBrakeEnabled.h
 return_getAutoBrakeEnabled.cpp
-return_getDesiredVelocityOne.h
-return_getDesiredVelocityOne.cpp
-return_getDesiredVelocityAll.h
-return_getDesiredVelocityAll.cpp
-return_getDesiredVelocityGroup.h
-return_getDesiredVelocityGroup.cpp
+return_getRefVelocityOne.h
+return_getRefVelocityOne.cpp
+return_getRefVelocityAll.h
+return_getRefVelocityAll.cpp
+return_getRefVelocityGroup.h
+return_getRefVelocityGroup.cpp
 return_getAxes.h
 return_getAxes.cpp
 return_getPid.h
