+++ src/devices/messages/ControlBoardMsgs/idl_generated_code/return_getRefVelocityOne.h
@@ -8,27 +8,27 @@
 // This is an automatically generated file.
 // It could get re-generated if the ALLOW_IDL_GENERATION flag is on.
 
-#ifndef YARP_THRIFT_GENERATOR_STRUCT_RETURN_getRefVelocITYALL_H
-#define YARP_THRIFT_GENERATOR_STRUCT_RETURN_getRefVelocITYALL_H
+#ifndef YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETREFVELOCITYONE_H
+#define YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETREFVELOCITYONE_H
 
 #include <yarp/os/Wire.h>
 #include <yarp/os/idl/WireTypes.h>
 #include <yarp/dev/ReturnValue.h>
 
-class return_getRefVelocityAll :
+class return_getRefVelocityOne :
         public yarp::os::idl::WirePortable
 {
 public:
     // Fields
     yarp::dev::ReturnValue ret{};
-    std::vector<double> vel{};
+    double vel{0.0};
 
     // Default constructor
-    return_getRefVelocityAll() = default;
+    return_getRefVelocityOne() = default;
 
     // Constructor with field values
-    return_getRefVelocityAll(const yarp::dev::ReturnValue& ret,
-                                 const std::vector<double>& vel);
+    return_getRefVelocityOne(const yarp::dev::ReturnValue& ret,
+                             const double vel);
 
     // Read structure on a Wire
     bool read(yarp::os::idl::WireReader& reader) override;
@@ -46,7 +46,7 @@ public:
     std::string toString() const;
 
     // If you want to serialize this class without nesting, use this helper
-    typedef yarp::os::idl::Unwrapped<return_getRefVelocityAll> unwrapped;
+    typedef yarp::os::idl::Unwrapped<return_getRefVelocityOne> unwrapped;
 
 private:
     // read/write ret field
@@ -62,4 +62,4 @@ private:
     bool nested_write_vel(const yarp::os::idl::WireWriter& writer) const;
 };
 
-#endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_getRefVelocITYALL_H
+#endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETREFVELOCITYONE_H
